[analyzer WIP] Consolidate array bound checkers
DO NOT MERGE THIS YET!
The ArrayBoundV2 checker still produces too many false positives because
the information (known constraints and values) provided by the analyzer
is frequently misleading (especially around situations that involve
loops).

itself is reasonably stable, but unfortunately
there are many deficiencies of the integer modeling logic of the Clang
SA engine, so

Before this commit, there were two alpha checkers that used different
algorithms/logic for detecting out of bounds memory access: the old
`alpha.security.ArrayBound` and the experimental, more complex
`alpha.security.ArrayBoundV2`.

After lots of quality improvement commits ArrayBoundV2 is now stable
enough to be moved out of the alpha stage. (It is not perfect: I know
about several corner cases where it produces false negatives or somewhat
vague error messages; but it reaches the average quality of non-alpha
checkers and it would be helpful for the users.)

At this point `alpha.security.ArrayBound` became obsolete (there is a
better tool for the same task), so I'm removing it from the codebase.
With this I can eliminate the ugly "V2" version mark almost everywhere
and rename `alpha.security.ArrayBoundV2` to `security.ArrayBound`.

(The version mark is preserved in the filename "ArrayBoundCheckerV2",
to ensure a clear git history. I'll rename it to "ArrayBoundChecker.cpp"
in a separate commit.)

This commit adapts the unit tests of `alpha.security.ArrayBound` to
testing the new `security.ArrayBound` (= old ArrayBoundV2). Currently
the names of the test files are very haphazard, I'll probably create a
separate followup commit that consolidates this.