[analyzer WIP] Proof-of-concept bad scaled pointer arithmetic checker
This commit implements alpha.security.BadScaledPointerArithmetic, a
checker for detecting suspicious code where the result of a `sizeof` or
`offsetof` operator is added to a non-char pointer. (Pointer arithmetic
automatically scales its integer operand by the size of the pointed
type; if the size of the pointed type is >1 bit, then applying this to a
size or offset value is practically always a bug.)

This checker is alpha quality because:
(1) The bug reports are vague.
(2) The implementation adds a new data member to `SVal` because it needs
to track the origin of concrete integers (while the existing
infrastructure is only suitable for tracking unknown, symbolic values).

Co-authored-by: Zoltan Porkoláb <zoltan.porkolab@ericsson.com>