[EXPERIMENTAL][analyzer] Limit loop modeling to two iterations
DO NOT MERGE; this is just a proof-of-concept patch.

Because it is justified to assume that a loop can run for two iterations
(otherwise the user wouldn't have written a loop), but it's not
justified to assume that the loop condition can be satisfied three times
(because there are examples where e.g. the developer knows that a data
structure may have <= 2 elements, and then the analyzer shouldn't
"hallucinate" an execution path where three iterations are executed).

Moreover, it's unlikely that a bug will be detectable in the third
iteration (after two iterations where it didn't appear), so it's
plausible that this could become the new default behavior of CSA.