[analyzer][EXPERIMENTAL] Widen instead of ambiguous loop exits
This is a proof of concept change, do not merge!

This commit does two things:
(1) it hardcodes the activation of the loop widening feature;
(2) it prevents the "exit the loop" branches in situation where the
value of the loop condition is ambiguous.

This is a rough and inaccurate approximation of my planned loop modeling
behavior.

Note that the tests are not adapted to this change and some of them are
failing.