[analyzer] Don't assume third iteration in loops
This commit ensures that if the loop condition is opaque (the analyzer
cannot determine whether it's true or false) and there were at least two
iterations, then the analysis doesn't enter a third iteration.

This heuristic is introduced to eliminate some false positives where
loops iterate over two-element data structures, but the analyzer doesn't
understand the loop condition and assumes that it may peform a third
iteration. (This produced 100+ false positive out-of-bounds reports in
the FFMPEG codebase and several other unwanted reports there and
in other open source projects).

Moreover, this will provide performance improvements in the sense that
the analyzer won't waste time on traversing the execution paths with 3
or 4 iterations in a loop (which are very similar to the paths with 2
iterations) and therefore will be able to traverse more branches
elsewhere on the ExplodedGraph.

This heuristic is disabled if the user enables the widen-loops analyzer
option (which is disabled by default), because the "simulate one final
iteration after the invalidation" execution path would be suppressed by
the "exit the loop if the loop condition is opaque and there were at
least two iterations" logic. If we want to support loop widening, we
would need to ensure that it is compatible with this heuristic (this
could belong to a follow-up commit).