[analyzer][EXPERIMENTAL] Only suppress branches assuming skipped loops
This is an intermediate version to separate the effects of the
suppression graph (performed by this revision) and the perturbations of
the exploded graph that are caused when we don't (wastefully) traverse
the execution paths that we're going to discard.

Note that this is an experimental commit for measurement purposes, the
inline comments are not updated to reflect the changes in the code.

Do not merge this to the main branch!